Refactor reader testing into a harness
