+++ sdk/metric/manual_reader_test.go
@@ -15,60 +15,9 @@
 package metric // import "go.opentelemetry.io/otel/sdk/metric/reader"
 
 import (
-	"context"
 	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"go.opentelemetry.io/otel/sdk/metric/export"
 )
 
-func TestManualReaderNotRegistered(t *testing.T) {
-	rdr := &manualReader{}
-
-	_, err := rdr.Collect(context.Background())
-	require.ErrorIs(t, err, ErrReaderNotRegistered)
-}
-
-type testProducer struct{}
-
-var testMetrics = export.Metrics{
-	// TODO: test with actual data.
-}
-
-func (p testProducer) produce(context.Context) (export.Metrics, error) {
-	return testMetrics, nil
-}
-
-func TestManualReaderProducer(t *testing.T) {
-	rdr := &manualReader{}
-	rdr.register(testProducer{})
-
-	m, err := rdr.Collect(context.Background())
-	assert.NoError(t, err)
-	assert.Equal(t, testMetrics, m)
-}
-
-func TestManualReaderCollectAfterShutdown(t *testing.T) {
-	rdr := &manualReader{}
-	rdr.register(testProducer{})
-	err := rdr.Shutdown(context.Background())
-	require.NoError(t, err)
-
-	m, err := rdr.Collect(context.Background())
-	assert.ErrorIs(t, err, ErrReaderShutdown)
-	assert.Equal(t, export.Metrics{}, m)
-}
-
-func TestManualReaderShutdown(t *testing.T) {
-	rdr := &manualReader{}
-	rdr.register(testProducer{})
-
-	err := rdr.Shutdown(context.Background())
-	require.NoError(t, err)
-
-	err = rdr.Shutdown(context.Background())
-	assert.ErrorIs(t, err, ErrReaderShutdown)
-
+func TestManualReader(t *testing.T) {
+	testReaderHarness(t, func() Reader { return NewManualReader() })
 }
